Rename location to locationHelpers
